requirements.txt
chainlit
pydantic
llama-index
llama-index-llms-openai
llama-index-agent-openai